requirements.txt
MySQL-python==1.2.4
PyYAML==3.10
SQLAlchemy==0.7.10
Werkzeug==0.9.4
argparse==1.2.1
distribute==0.6.34
sae-python-dev==1.2.0
tornado==2.4.1
wsgiref==0.1.2
